Neue Methode: - save Spieler, sodass der Controller nicht auf den Repo direkt zugreift

+++ src/main/java/dhsn/verwaltung/spieler/service/SpielerService.java
@@ -10,9 +10,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import dhsn.verwaltung.spieler.model.domain.Spieler;
-import dhsn.verwaltung.spieler.model.domain.DTO.SpielerRegisterDTO;
-import dhsn.verwaltung.spieler.model.domain.DTO.SpielerUpdateDTO;
+import dhsn.verwaltung.spieler.model.domain.SpielerDTO.SpielerRegisterDTO;
+import dhsn.verwaltung.spieler.model.domain.SpielerDTO.SpielerUpdateDTO;
 import dhsn.verwaltung.spieler.model.identity.Benutzer;
+import dhsn.verwaltung.spieler.model.identity.Role;
 import dhsn.verwaltung.spieler.repository.SpielerRepository;
 import jakarta.transaction.Transactional;
 
@@ -26,6 +27,10 @@ public class SpielerService {
   ) {
     this.spielerRepo = sr;
     this.passwordEncoder = passwordEncoder;
+  }
+
+  public Spieler getSpieler(Long id) {
+    return spielerRepo.findById(id).orElseThrow(()->new UsernameNotFoundException("ID ist unbekannt: " +id));
   }
 
   public SpielerUpdateDTO getBasicSpielerDTO(Long id) {
@@ -51,11 +56,12 @@ public class SpielerService {
     return spielerRepo.findAll();
   }
 
+  @Transactional
   public Benutzer register(SpielerRegisterDTO spielerDTO) {
     Spieler neuerSpieler = new Spieler(spielerDTO.getUsername(),
-    passwordEncoder.encode(spielerDTO.getPasswort()), spielerDTO.getRole(),
+    passwordEncoder.encode(spielerDTO.getPasswort()), Role.ROLE_SPIELER,
     spielerDTO.getVorname(), spielerDTO.getNachname(), spielerDTO.getGeburtsdatum(),
-    spielerDTO.getPosition(), spielerDTO.getRueckennummer());
+    spielerDTO.getPosition());
     spielerRepo.save(neuerSpieler);
     return neuerSpieler;
   }
@@ -74,30 +80,35 @@ public class SpielerService {
     datenBankSpieler.setPosition(spielerBDTO.getPosition());
     datenBankSpieler.setRueckennummer(spielerBDTO.getRueckennummer());
 
-    if (
-      !spielerBDTO.getPasswort().isEmpty() &&
-      spielerBDTO.getPasswort() != null 
-      && !datenBankSpieler.getPasswort().equals(passwordEncoder.encode(spielerBDTO.getPasswort()))
-    ) {
-        
+    if (!spielerBDTO.getPasswort().isEmpty() && spielerBDTO.getPasswort() != null) {
       datenBankSpieler.setPasswort(passwordEncoder.encode(spielerBDTO.getPasswort()));
     }
     spielerRepo.save(datenBankSpieler);
   }
 
   @Transactional
-  public void deleteSpieler(Long id) {
+  public void deleteSpieler(Long id) throws ResponseStatusException {
     //Test ob ID in DB vorhanden ist
-    spielerRepo.findById(id).orElseThrow(() -> 
-    new ResponseStatusException(HttpStatus.NOT_FOUND, "ID: "+id+" unbekannt"));
+    if(!spielerRepo.existsById(id)){
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND, "ID: "+id+" unbekannt");
+    }
     spielerRepo.deleteById(id);
   }
 
+  @Transactional
   public void setGehaltErhoung(Long id, boolean willErhoung) {
     Spieler spieler = spielerRepo.findById(id).orElseThrow(()->
       new ResponseStatusException(HttpStatus.NOT_FOUND, "ID: "+id+" unbekannt")
     );
     spieler.setWillGehaltsErhoeung(willErhoung);
     spielerRepo.save(spieler);
+  }
+
+  public void save(Spieler spieler) {
+    spielerRepo.save(spieler);
+  }
+
+  public List<Spieler> getFreieSpieler() {
+    return spielerRepo.findByVereinIsNull();
   }
 }
